Use http instead of https to allow travis fetch rack-livereload gem
This was printed out by Travis CI:

Could not verify the SSL certificate for
https://s3.amazonaws.com/production.s3.rubygems.org/quick/Marshal.4.8/rack-livereload-0.3.15.gemspec.rz.
There is a chance you are experiencing a man-in-the-middle attack, but most
likely your system doesn't have the CA certificates needed for verification. For
information about OpenSSL certificates, see bit.ly/ruby-ssl. To connect without
using SSL, edit your Gemfile sources and change 'https' to 'http'.