Add user traits donald and scrooge
We generate all infos in factories automatically, e.g. names (but not passwords). To make specs more specific, we want to test on strings, not on generated data, so we need some standard traits of which we know what to test. For normal users, we will use donald, for admin users we will use scrooge.